Get rid of top level functions and move tests accordingly
