+++ rust/stracciatella/src/config/ja2_json.rs
@@ -32,14 +32,14 @@ pub struct Ja2Json {
     path: PathBuf,
 }
 
-fn build_json_config_location(stracciatella_home: &PathBuf) -> PathBuf {
-    resolve_existing_components(&Path::new("ja2.json"), Some(&stracciatella_home), true)
+fn build_json_config_location(stracciatella_home: &Path) -> PathBuf {
+    resolve_existing_components(&Path::new("ja2.json"), Some(stracciatella_home.as_ref()), true)
 }
 
 impl Ja2Json {
     /// Construct a Ja2Json instance from the stracciatella home directory
-    pub fn from_stracciatella_home(stracciatella_home: &PathBuf) -> Self {
-        let path = build_json_config_location(stracciatella_home);
+    pub fn from_stracciatella_home<P>(stracciatella_home: P) -> Self where P: AsRef<Path> {
+        let path = build_json_config_location(stracciatella_home.as_ref());
         Ja2Json { path }
     }
 
@@ -159,5 +159,246 @@ impl Ja2Json {
         }
 
         Ok(())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::io::prelude::*;
+    use std::path::Path;
+
+    use tempfile::TempDir;
+
+    use crate::config::VanillaVersion;
+    use crate::fs;
+    use crate::fs::File;
+    use super::*;
+
+    pub fn write_temp_folder_with_ja2_json(contents: &[u8]) -> TempDir {
+        let dir = TempDir::new().unwrap();
+        let ja2_home_dir = dir.path().join(".ja2");
+        let file_path = ja2_home_dir.join("ja2.json");
+
+        fs::create_dir(ja2_home_dir).unwrap();
+        std::fs::write(file_path, contents).unwrap();
+
+        dir
+    }
+
+    #[test]
+    fn ensure_existence_should_ensure_existence_of_config_dir() {
+        let dir = TempDir::new().unwrap();
+        let home_path = dir.path().join("ja2_home");
+        let expected_path = home_path.join("ja2.json");
+
+        let ja2json = Ja2Json::from_stracciatella_home(&home_path);
+
+        ja2json.ensure_existence().unwrap();
+
+        assert!(home_path.exists());
+        assert!(expected_path.is_file());
+    }
+
+    #[test]
+    fn ensure_existence_should_not_overwrite_existing_ja2json() {
+        let dir = write_temp_folder_with_ja2_json(b"Test");
+        let ja2json = Ja2Json::from_stracciatella_home(dir.path().join(".ja2"));
+        let ja2json_path = dir.path().join(".ja2/ja2.json");
+
+        ja2json.ensure_existence().unwrap();
+
+        let mut f = File::open(ja2json_path.clone()).unwrap();
+        let mut content: Vec<u8> = vec![];
+        f.read_to_end(&mut content).unwrap();
+
+        assert!(ja2json_path.is_file());
+        assert_eq!(content, b"Test");
+    }
+
+    #[test]
+    fn apply_to_engine_options_should_fail_with_missing_file() {
+        let mut engine_options = EngineOptions::default();
+        let temp_dir = TempDir::new().unwrap();
+        let ja2json = Ja2Json::from_stracciatella_home(temp_dir.path().join(".ja2"));
+        let result = ja2json.apply_to_engine_options(&mut engine_options);
+
+        assert!(result.is_err());
+        assert!(
+            result
+                .err()
+                .unwrap()
+                .starts_with("Error reading ja2.json config file:"),
+            "error didn't start with correct string"
+        );
+    }
+
+    #[test]
+    fn apply_to_engine_options_should_fail_with_invalid_json() {
+        let mut engine_options = EngineOptions::default();
+        let temp_dir = write_temp_folder_with_ja2_json(b"{ not json }");
+        let ja2json = Ja2Json::from_stracciatella_home(temp_dir.path().join(".ja2"));
+
+        assert_eq!(
+            ja2json.apply_to_engine_options(&mut engine_options),
+            Err(String::from(
+                "Error parsing ja2.json config file: key must be a string at line 1 column 3"
+            ))
+        );
+    }
+
+    #[test]
+    fn apply_to_engine_options_should_be_able_to_change_game_dir() {
+        let mut engine_options = EngineOptions::default();
+        let temp_dir = write_temp_folder_with_ja2_json(b"{ \"game_dir\": \"/dd\" }");
+        let ja2json = Ja2Json::from_stracciatella_home(temp_dir.path().join(".ja2"));
+
+        ja2json.apply_to_engine_options(&mut engine_options).unwrap();
+
+        assert_eq!(engine_options.vanilla_game_dir, Path::new("/dd"));
+    }
+
+    #[test]
+    fn apply_to_engine_options_should_be_able_to_change_game_dir_with_data_dir() {
+        let mut engine_options = EngineOptions::default();
+        // data_dir is an alias to game_dir
+        let temp_dir = write_temp_folder_with_ja2_json(b"{ \"data_dir\": \"/dd\" }");
+        let ja2json = Ja2Json::from_stracciatella_home(temp_dir.path().join(".ja2"));
+
+        ja2json.apply_to_engine_options(&mut engine_options).unwrap();
+
+        assert_eq!(engine_options.vanilla_game_dir, Path::new("/dd"));
+    }
+
+    #[test]
+    fn apply_to_engine_options_should_be_able_to_change_fullscreen_value() {
+        let mut engine_options = EngineOptions::default();
+        let temp_dir = write_temp_folder_with_ja2_json(b"{ \"fullscreen\": true }");
+        let ja2json = Ja2Json::from_stracciatella_home(temp_dir.path().join(".ja2"));
+
+        ja2json.apply_to_engine_options(&mut engine_options).unwrap();
+
+        assert_eq!(engine_options.start_in_fullscreen, true);
+    }
+
+    #[test]
+    fn apply_to_engine_options_should_be_able_to_change_debug_value() {
+        let mut engine_options = EngineOptions::default();
+        let temp_dir = write_temp_folder_with_ja2_json(b"{ \"debug\": true }");
+        let ja2json = Ja2Json::from_stracciatella_home(temp_dir.path().join(".ja2"));
+
+        ja2json.apply_to_engine_options(&mut engine_options).unwrap();
+
+        assert_eq!(engine_options.start_in_debug_mode, true);
+    }
+
+    #[test]
+    fn apply_to_engine_options_should_be_able_to_start_without_sound() {
+        let mut engine_options = EngineOptions::default();
+        let temp_dir = write_temp_folder_with_ja2_json(b"{ \"nosound\": true }");
+        let ja2json = Ja2Json::from_stracciatella_home(temp_dir.path().join(".ja2"));
+
+        ja2json.apply_to_engine_options(&mut engine_options).unwrap();
+
+        assert_eq!(engine_options.start_without_sound, true);
+    }
+
+    #[test]
+    fn apply_to_engine_options_should_not_be_able_to_run_help() {
+        let mut engine_options = EngineOptions::default();
+        let temp_dir = write_temp_folder_with_ja2_json(b"{ \"help\": true, \"show_help\": true }");
+        let ja2json = Ja2Json::from_stracciatella_home(temp_dir.path().join(".ja2"));
+
+        ja2json.apply_to_engine_options(&mut engine_options).unwrap();
+
+        assert_eq!(engine_options.show_help, false);
+    }
+
+    #[test]
+    fn apply_to_engine_options_should_not_be_able_to_run_unittests() {
+        let mut engine_options = EngineOptions::default();
+        let temp_dir =
+            write_temp_folder_with_ja2_json(b"{ \"unittests\": true, \"run_unittests\": true }");
+        let ja2json = Ja2Json::from_stracciatella_home(temp_dir.path().join(".ja2"));
+
+        ja2json.apply_to_engine_options(&mut engine_options).unwrap();
+
+        assert_eq!(engine_options.run_unittests, false);
+    }
+
+    #[test]
+    fn apply_to_engine_options_should_not_be_able_to_run_editor() {
+        let mut engine_options = EngineOptions::default();
+        let temp_dir =
+            write_temp_folder_with_ja2_json(b"{ \"editor\": true, \"run_editor\": true }");
+        let ja2json = Ja2Json::from_stracciatella_home(temp_dir.path().join(".ja2"));
+
+        ja2json.apply_to_engine_options(&mut engine_options).unwrap();
+
+        assert_eq!(engine_options.run_editor, false);
+    }
+
+    #[test]
+    fn apply_to_engine_options_should_fail_with_invalid_mod() {
+        let mut engine_options = EngineOptions::default();
+        let temp_dir = write_temp_folder_with_ja2_json(b"{ \"mods\": [ \"a\", true ] }");
+        let ja2json = Ja2Json::from_stracciatella_home(temp_dir.path().join(".ja2"));
+
+        assert_eq!(ja2json.apply_to_engine_options(&mut engine_options), Err(String::from("Error parsing ja2.json config file: invalid type: boolean `true`, expected a string at line 1 column 21")));
+    }
+
+    #[test]
+    fn apply_to_engine_options_should_continue_with_multiple_known_switches() {
+        let mut engine_options = EngineOptions::default();
+        let temp_dir =
+            write_temp_folder_with_ja2_json(b"{ \"debug\": true, \"mods\": [ \"m1\", \"a2\" ] }");
+        let ja2json = Ja2Json::from_stracciatella_home(temp_dir.path().join(".ja2"));
+
+        ja2json.apply_to_engine_options(&mut engine_options).unwrap();
+
+        assert_eq!(engine_options.start_in_debug_mode, true);
+        assert_eq!(engine_options.mods.len(), 2);
+    }
+
+    #[test]
+    fn apply_to_engine_options_should_fail_with_unknown_resversion() {
+        let mut engine_options = EngineOptions::default();
+        let temp_dir = write_temp_folder_with_ja2_json(b"{ \"resversion\": \"TESTUNKNOWN\" }");
+        let ja2json = Ja2Json::from_stracciatella_home(temp_dir.path().join(".ja2"));
+
+        assert_eq!(ja2json.apply_to_engine_options(&mut engine_options), Err(String::from("Error parsing ja2.json config file: unknown variant `TESTUNKNOWN`, expected one of `DUTCH`, `ENGLISH`, `FRENCH`, `GERMAN`, `ITALIAN`, `POLISH`, `RUSSIAN`, `RUSSIAN_GOLD`, `SIMPLIFIED_CHINESE` at line 1 column 29")));
+    }
+
+    #[test]
+    fn apply_to_engine_options_should_return_the_correct_resversion_for_russian() {
+        let mut engine_options = EngineOptions::default();
+        let temp_dir = write_temp_folder_with_ja2_json(b"{ \"resversion\": \"RUSSIAN\" }");
+        let ja2json = Ja2Json::from_stracciatella_home(temp_dir.path().join(".ja2"));
+
+        ja2json.apply_to_engine_options(&mut engine_options).unwrap();
+
+        assert_eq!(engine_options.resource_version, VanillaVersion::RUSSIAN);
+    }
+
+    #[test]
+    fn apply_to_engine_options_should_return_the_correct_resversion_for_italian() {
+        let mut engine_options = EngineOptions::default();
+        let temp_dir = write_temp_folder_with_ja2_json(b"{ \"resversion\": \"ITALIAN\" }");
+        let ja2json = Ja2Json::from_stracciatella_home(temp_dir.path().join(".ja2"));
+
+        ja2json.apply_to_engine_options(&mut engine_options).unwrap();
+
+        assert_eq!(engine_options.resource_version, VanillaVersion::ITALIAN);
+    }
+
+    #[test]
+    fn apply_to_engine_options_should_return_the_correct_resolution() {
+        let mut engine_options = EngineOptions::default();
+        let temp_dir = write_temp_folder_with_ja2_json(b"{ \"res\": \"1024x768\" }");
+        let ja2json = Ja2Json::from_stracciatella_home(temp_dir.path().join(".ja2"));
+
+        ja2json.apply_to_engine_options(&mut engine_options).unwrap();
+
+        assert_eq!(engine_options.resolution.0, 1024);
+        assert_eq!(engine_options.resolution.1, 768);
     }
 }
